+++ src/ExtendedPdoInterface.php
@@ -30,6 +30,15 @@ interface ExtendedPdoInterface extends PdoInterface
 
     /**
      *
+     * Manual method that disconnects PDO and unsets PDO object
+     *
+     * @return null
+     *
+     */
+    public function disconnect();
+
+    /**
+     *
      * Performs a statement and returns the number of affected rows.
      *
      * @param string $statement The SQL statement to prepare and execute.
@@ -212,7 +221,7 @@ interface ExtendedPdoInterface extends PdoInterface
      *
      * Returns the underlying PDO connection object.
      *
-     * @return PDO
+     * @return PDO or Null if connection was manually disconnected
      *
      */
     public function getPdo();